[MSLITE] Codex clean

+++ mindspore/lite/src/tensor.cc
@@ -257,7 +257,7 @@ size_t Tensor::Size() const {
     MS_LOG(INFO) << "Element number of tensor should large than 0 : " << element_num << ", shape: " << shape_;
     return 0;
   }
-  return element_size * element_num;
+  return element_size * static_cast<size_t>(element_num);
 }
 
 int64_t Tensor::ElementsNum() const {
